requirements.txt
# Document Loading and Processing Requirements
tqdm==4.67.1
tiktoken==0.9.0
qdrant-client==1.14.3
python-dotenv==1.1.0
openai==1.90.0
langchain
langchain-qdrant
langchain[openai]